// Meteo-Alert/src/models/types/meteoData.ts
type meteoData = {
    neige: number;
    pluie: number;
    humidite: number;
    visibilite: number;
    ressenti: number;
    temperature: number;
    tempMin: number;
    tempMax: number;
    ventRafale: number;
    ventVitesse: number;
    ventDirection: number;
    pression: number;
    pressionTerre: number;
    pressionMer: number;
    nuage: number
}

export default meteoData;